Add documentation for Wayland
